+++ src/main/java/org/devcourse/springbasic/domain/voucher/dao/VoucherJdbcRepository.java
@@ -3,7 +3,6 @@ package org.devcourse.springbasic.domain.voucher.dao;
 import lombok.RequiredArgsConstructor;
 import org.devcourse.springbasic.domain.voucher.domain.Voucher;
 import org.devcourse.springbasic.domain.voucher.domain.VoucherType;
-import org.devcourse.springbasic.domain.voucher.dto.VoucherDto;
 import org.springframework.context.annotation.Profile;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.dao.EmptyResultDataAccessException;
@@ -15,7 +14,6 @@ import org.springframework.stereotype.Repository;
 import java.nio.ByteBuffer;
 import java.util.List;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 @Profile("dev")
 @Repository
@@ -25,45 +23,38 @@ public class VoucherJdbcRepository implements VoucherRepository {
     private final NamedParameterJdbcTemplate jdbcTemplate;
 
     @Override
-    public VoucherDto.ResponseDto findById(UUID voucherId) {
+    public Voucher findById(UUID voucherId) {
 
         String QUERY = "select * from vouchers where voucher_id = UUID_TO_BIN(:voucherId)";
         MapSqlParameterSource paramMap = new MapSqlParameterSource()
                 .addValue("voucherId", voucherId.toString().getBytes());
 
         try {
-            Voucher voucher = jdbcTemplate.queryForObject(
+            return jdbcTemplate.queryForObject(
                     QUERY,
                     paramMap,
                     VOUCHER_ROW_MAPPER
             );
-            return new VoucherDto.ResponseDto(voucher.getVoucherId(), voucher.getVoucherType(), voucher.getDiscountRate());
         } catch (EmptyResultDataAccessException emptyResultDataAccessException) {
             throw new IllegalArgumentException("해당 ID인 바우처가 없습니다.");
         }
     }
 
     @Override
-    public List<VoucherDto.ResponseDto> findAll() {
+    public List<Voucher> findAll() {
         String QUERY = "select * from vouchers";
-        List<Voucher> vouchers = jdbcTemplate.query(QUERY, VOUCHER_ROW_MAPPER);
-        return vouchers.stream()
-                .map(voucher -> new VoucherDto.ResponseDto(voucher.getVoucherId(), voucher.getVoucherType(), voucher.getDiscountRate()))
-                .collect(Collectors.toList());
+        return jdbcTemplate.query(QUERY, VOUCHER_ROW_MAPPER);
     }
 
     @Override
-    public UUID save(VoucherDto.SaveRequestDto voucherDto) {
+    public UUID save(Voucher voucher) {
 
         String QUERY = "insert into vouchers(voucher_id, discount_amount, voucher_type)" +
                      "values (UUID_TO_BIN(:voucherId), :discountAmount, :voucherType)";
-        VoucherType voucherType = voucherDto.getVoucherType();
-        Voucher voucher = voucherType.getVoucherSupplier().get();
-
         MapSqlParameterSource paramMap = new MapSqlParameterSource()
-                .addValue("voucherType", voucher.getVoucherType())
-                .addValue("voucherId", voucher.getVoucherId())
-                .addValue("discountAmount", voucher.getDiscountRate());
+                .addValue("voucherId", voucher.getVoucherId().toString().getBytes())
+                .addValue("discountAmount", voucher.getDiscountRate())
+                .addValue("voucherType", voucher.getVoucherType().toString());
 
         try {
             jdbcTemplate.update(QUERY, paramMap);
