+++ src/main/java/org/devcourse/springbasic/domain/voucher/dao/VoucherMemoryRepository.java
@@ -1,14 +1,14 @@
 package org.devcourse.springbasic.domain.voucher.dao;
 
 import org.devcourse.springbasic.domain.voucher.domain.Voucher;
-import org.devcourse.springbasic.domain.voucher.domain.VoucherType;
-import org.devcourse.springbasic.domain.voucher.dto.VoucherDto;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Repository;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.stream.Collectors;
 
 @Profile("local")
 @Repository
@@ -17,25 +17,22 @@ public class VoucherMemoryRepository implements VoucherRepository {
     private final Map<UUID, Voucher> storage = new ConcurrentHashMap<>();
 
     @Override
-    public VoucherDto.ResponseDto findById(UUID voucherId) {
-        Optional<Voucher> voucher = Optional.ofNullable(storage.get(voucherId));
-        return voucher.map(v -> new VoucherDto.ResponseDto(v.getVoucherId(), v.getVoucherType(), v.getDiscountRate()))
-                .orElseThrow(() -> new IllegalArgumentException("해당 아이디인 바우처가 존재하지 않습니다."));
+    public Voucher findById(UUID voucherId) {
+        try {
+            return storage.get(voucherId);
+        } catch (NullPointerException nullPointerException) {
+            throw new IllegalArgumentException("해당 ID를 가진 회원이 없습니다.");
+        }
     }
 
     @Override
-    public List<VoucherDto.ResponseDto> findAll() {
-        return storage.values().stream()
-                .map(voucher -> new VoucherDto.ResponseDto(voucher.getVoucherId(), voucher.getVoucherType(), voucher.getDiscountRate()))
-                .collect(Collectors.toList());
+    public List<Voucher> findAll() {
+        return new ArrayList<>(storage.values());
     }
 
     @Override
-    public UUID save(VoucherDto.SaveRequestDto voucherDto) {
-        VoucherType voucherType = voucherDto.getVoucherType();
-        Voucher voucher = voucherType.getVoucherSupplier().get();
+    public UUID save(Voucher voucher) {
         storage.put(voucher.getVoucherId(), voucher);
-
         return voucher.getVoucherId();
     }
 }
