+++ src/main/java/org/devcourse/springbasic/domain/voucher/dao/VoucherRepository.java
@@ -1,12 +1,12 @@
 package org.devcourse.springbasic.domain.voucher.dao;
 
-import org.devcourse.springbasic.domain.voucher.dto.VoucherDto;
+import org.devcourse.springbasic.domain.voucher.domain.Voucher;
 
 import java.util.List;
 import java.util.UUID;
 
 public interface VoucherRepository {
-    VoucherDto.ResponseDto findById(UUID voucherId);
-    List<VoucherDto.ResponseDto> findAll();
-    UUID save(VoucherDto.SaveRequestDto voucherDto);
+    Voucher findById(UUID voucherId);
+    List<Voucher> findAll();
+    UUID save(Voucher voucher);
 }
